Move controller's root view to VideoControllerView
