+++ app/src/main/res/layout/activity_action_bar.xml
@@ -8,6 +8,6 @@
         android:id="@+id/fullscreenVideoView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        app:play_drawable="@drawable/ic_launcher_background"
-        app:pause_drawable="@drawable/ic_pause_white_48dp"/>
+        app:pause_drawable="@drawable/ic_pause_white_48dp"
+        app:play_drawable="@drawable/ic_launcher_background" />
 </LinearLayout>
